Temporary commit to be able to revert to a certain commit.

+++ libft/ft_strlen.c
@@ -3,17 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jgirard- <jgirard-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hdelmann <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 17:28:13 by jgirard-          #+#    #+#             */
-/*   Updated: 2022/03/24 23:13:43 by jgirard-         ###   ########.fr       */
+/*   Updated: 2023/01/12 14:11:29 by hdelmann         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"libft.h"
-#include <stdio.h>
 
-size_t	ft_strlen(const char	*str)
+int	ft_strlen(const char	*str)
 {
 	int	i;
